remove get_logger and PROJ_ROOT
